REFACTOR: move component to related folder
