+++ app/[locale]/Content.tsx
