+++ components/index.ts
@@ -1,8 +1,6 @@
-export { default as Brand } from './Brand';
 export { default as Button } from './Button';
 export { default as Icon } from './Icon';
 export { default as Card } from './Card';
 export { default as ContentLabels } from './ContentLabels';
-export { default as Content } from './Content';
 export { default as Slider } from './Slider';
 export { default as Popover } from './Popover';
